Move workflow config to instrument conf file
